Simulation class is the context in which the model is run [#95934926]
An instance of this class is created with simulation options.
Its a short-lived object created for each run.

Following a DCI pattern, the Simulation modifies the participating nodes,
and binds them to an integration function, and initializes there state.  We might support
different integration methods in the future. (current implemtation is a simple forward Euler)

The only context in which it makes sense for nodes to have state history,
and an integration function is when they are being used in a simulation. This is an
attempt to keep all the simulation code close together.

https://www.pivotaltracker.com/story/show/95934926