Return success of undo/redo to CODAP
If an undo or a redo fails (specifically, if any of the listers
on an action return a failure), we can inform CODAP of this,
which at the moment will simply cause it to clear its history
stack to prevent getting into a confused state. (This is the same
behavior CODAP has if one of its own undoable actions fails to
undo or redo).