Fix mocha test broken by inclusion of codap-connect
With codap-connect now being required directly by the undo-redo manager,
codap-connect's reference to iFramePhone was also being brought into the
test, and iFramePhone has a dependency on `window`.

This fixes the test by both having an empty global object `window`, and
also by stubbing out the contents of codap-connect. (Just doing the latter
wasn't enough, as codap-connect's `require` statement is still processed
even when it's being stubbed by Sinon.)

This also reverts the createAndExecuteCommand function to return the result
of `execute`, as it used it, as this is checked by the test.