Add ability to batch undo/redo commands; batch replace palette item
When we replace one palette item with another, we perform
one action for each node replacement, and one action for
the deletion of the palette item.

This can result in a stack of several actions, which if the user tries
to undo will result in several clicks of undo (and similarly for redo).

This feature allows us to call `startCommandBatch` and `endCommandCatch`
before and after an arbitrarily-long set of commands that we would
like the user to be able to undo and redo with one click.

This is modeled slightly off SproutCore's start/endRunLoop.

Any nested calls to startCommandBatch will simply be added to the
current commandBatch that has been created.

[#101973812]