Use command batch when editing node titles
Prior to this, when node titles were edited, each letter would add a new
command to the undo history, which would require being individually
undone, one letter at a time.

This batches all the edits until the user has finished editing.

We call startCommandBatch with every edit to the title, but this all gets
folded into the existing batch if the user was already editing. (We could
alternatively call startCommandBatch just once, in node-view's startEditing
command, but then we'd also need another one to catch the initial editing
when a node is first created.)

This adds functions to the graph-store to avoid node-view needing to know
about the undoRedoManager.

[#101973812]