Add editing panel for defining node values with words
This adds a toggleable panel that allows a user to switch between defining
a node's value semi-quantitatively (low-high) or with numbers.

This adds a new property to a node, because any node that has been defined
in one way should keep being defined that way when the user re-opens the
editing panel.

The property `valueDefinedSemiQuantitatively` is a bit of a mouthful,
but it describes the intent. Note that this only describes how it's being
defined, nothing about the actual value changes.

[#97759284]