Edit min and max when properties clicked
This makes the min and max values in the quantitative view pop out
slightly to show they are editable. When they are clicked, they turn
into input views, which update min and max as they are edited.

The slider automatically does the right thing, as the entire view is
rerendered when min and max are updated.

Note that this does not yet include sanity checks on min, max and
initialValue, which will be handled by the node itself in the next commit.

[#99423710]