Ensure min, max and initialValue are sane
Each time a node's properties are changed, we ensure than the min is
less or equal to the max, and that the initialValue is between the two.

Because the view updates due to the properties being updated, this means
that as the min or max is changed, the other property, the initial value,
and the slider all instantly do the right thing.

[#99423710]