Check if graph is valid, throw error if not; +tests
Before we run a simulation, we check to see if a graph is valid. In this
case, we define "invalid" to be a graph with any loop of dependent variables
(i.e. nodes that (1) have incoming links, and (2) are not accumulators).

Because we run this check once at the start of the simulation, and the
simulation cannot change during its running (at least not yet), we get
to keep our fairly simple integration function, and we never have to worry
that our integration function will encounter a loop.

Currently we just throw an internal error. We don't yet show this to the user.
However, graphIsValid is a public method as we'll be calling it to set the
state of the run button next.

Add tests to ensure that invalid graphs throw errors.

[#101699204] [#101699240]