Check graph is valid on each change; update run button
The app-view mixin observes the graphStore, and on each change creates
a simulation and checks if it is valid (without running it).

If the graph is invalid, the document-actions-view makes the Run
button red.

If the button is clicked, an alert is shown. A future version can show
a nicely styled popup, and possibly highlight the loop in question,
although this may be more complicated.

[#101699240]