Don't show value editing for dependent variables
If a node is a dependent variable (has incoming links) and is not a
collector, then its initial value is meaningless and is ignored
by the simulation.

In this situation, we hide the value editing and inform the user that
it is a dependent variable.

Note that this only affects the view. The node still has an initial
value under the hood, because if a user sets a value of an independent
variable, then links to it (making it dependent), then breaks that link,
we want to recover the original value set.

[#101698452]