Create app settings panel, accessible from menu
We have a new App Settings panel, accessible from the existing (but
previously non-functional) "Advanced Settings" menu item, with just
one setting, "cap node values."

Add AppSettingsStore to handle the settings and the view state.

Currently, the Cap Values settings doesn't do anything, and the value
is not serialized, that will be added in the next two commits.

[#103222922]