Add settings gear icon when framed in CODAP
We don't have the menu bar, so a new gear icon after undo-redo will
open the same settings panel. Perhaps eventually this could be moved
to CODAP's own panel bar (like the other CODAP gear menus).

As part of this, remove ModalAppSettings view from globalNavView to
documentActionsView, which is available in both models.

[#103222922]