Limit node values to min-max range if capping values
This limits all nodes to their range bounds if the global setting
`capNodeValues` is true.

For accumulators, this works transparently.

For non-accumulating dependent nodes, it's a little odd as you can't
set min and max in view while node is dependent, but the bounds are
still applied. If you want to change the min and max, you have to make
the node independent (or change it to an accumulator) first.

[#103222922]