Remove value from Slider's state
This makes it more like a controlled component, such as `<input>`
(see [1](https://facebook.github.io/react/docs/forms.html#controlled-components)
and [2](http://leftdevel.com/blog/reactjs-controlled-forms/))

By not duplicating "value" in state, we have just one source of truth,
and remove any confusion of whether we're talking about the state or
the property at any point.

With a controlled component, we need to have some other variable owned
by the parent that gets updated, and then re-renders the slider component
with the new value. This seems like extra overhead, but in practice we'll
always want the slider to be attached to some property anyway. Updated
demo to reflect this.

Unfortunately, in this case it's not *exactly* the same as a React
`<input>` component, because the slider will actually move when you
drag it, even if you haven't hooked up the listener. This is because
the position of the handle when dragged is being set by jQuery, not the
value of `value`. But I think that's a minor concern.

Also change "min-height" to `minHeight` to remove one warning from JS console.