Make Slider min and max editable
Like `value`, min and max are controlled properties, and so a parent
needs to set min and max based on an onRangeChanged event.

Values are checked for sanity after user clicks away or hits return.

Most of this code from node-value-inspector-view, which will eventually
be replaced by this view.

Update slider demo.

[#102292942]