When we cancel Replace Item, don't add new image to undo stack
Previously, if we started to replace an item, added a new item as part
of this process, then hit Cancel in the Replace dialog, the newly-added
item would correctly disappear (now that the previous fix is in),
but, as an "easter egg," the user could re-add that item by hitting
the redo button.

This allows us to call `undo(true)` to undo and drop the action off
the stack, preventing it from being redone.

Because this is called programmatically, in a situation where an action
has just been performed, we can be certain we're always already at the
front of the undo stack, and so no redoable actions will be lost (i.e.
any potentially-redoable actions would have been cleared regardless).