Fix bug where blurring node name created empty action
There were two ways this bug could appear.

1. (More likely) after node creation, the title is editable. If you click
away, even to directly click on the Undo button, and empty action was
created by clicking way, so undo must be hit twice

2. If you click on the title for editing and then change your mind and
click away, an empty action is created.

This is fixed by

1. Checking if a node has actually changed before creating the undoable
change action

2. If we're in a batch mode, throw away any batch with no actions instead
of adding it to the undo stack.

NOTE: There is still a perhaps-surprising behavior that, if the node title
is open for editing (e.g. immediately when you drag it in), and you click
on the node and drag it around a bunch, all individual  drags and any title
changes are batched into one undoable action. This is because clicking on
the node and dragging it around don't blur the editable title. However,
since the *only* thing you can do without blurring the title is to drag
the node around, I think this is fine, and perhaps even more natural.