Remove extraneous info from simulation report
The initial value and the endState were only being used by the
old popup and in log.info. We can recreate the endState for log.info
without making it an explicit part of the report.

`duration` is the same as `steps` now, and each node doesn't need
to hold on to its own `time`, as this is being saved per frame.