Simulation publishes events, codap now listens to them
Instead of SimulationStore running a model, requesting a report, and
then explicitly passing that report to codap-connect, SimulationStore
now acts as a publisher of events, which anyone can listen to.

SimulationStore triggers simulationStart, simulationFramesGenerated
(with one or more frames), and simulationEnd.

The advantage of this is

1. Multiple listeners can be notified of simulation data, such as our
upcoming mini-graphs
2. Data from a simulation is modeled as asynchronous from the start,
which will make it possible to slow down the model and send individual
packages of data (next commit)

The simulation model now has the option of reporting frames as they
are generated, or bundling them all up and reporting them at the end.

A nice feature is that we can remove codap-connect completely from
simulation-store. If we could remove the remaining `require codap-connect`
statements from the rest of the app, it would make testing easier, as
we'd never need to deal with IframePhone in our tests...

Also update tests to now test for callbacks with the appropriate data.

[#102294388]