Add slow-speed simulations, wire up speed slider
Simulations can now run all at once (a synchronous loops which sends all
the values at the end), or at various slower speeds (an asynchronous
loop which sends 1 or more values at intervals).

The speed range is from 0 (2 steps/second) to 0.99 (30 steps/second) to
1 (instant).

The slower speeds use `requestAnimationFrame` to request control from
the processor each time the screen refreshes. It calculates the elapsed
time, and generates an appropriate number of frames for that time.

This means that if it sends a value to CODAP, and CODAP takes a significant
period of time before handing control back to `requestAnimationFrame`,
then the simulation will generate several frames to "catch up" with the
expected frame number.

So no matter how slowly CODAP or anything else (e.g. mini-graphs) takes
to render, a slowed-down simulation should complete within exactly one
refresh-paint of the expected end time (noting that that final refresh
paint may take a noticeable amount of time, but that's unavoidable).

The test shims `window.performance.now` and `requestAnimationTest`

[#102294388]