Update how we calculate speed:
- Speed 0: 1 step/second
- Speed 1: 3 steps/second, runs at least 5 seconds and at most 100 seconds
- Speed 2: 10 steps/second, runs at least 2.5 seconds and at most 50 seconds
- Speed 3: 20 steps/second, runs at least 1 seconds and at most 10 seconds
- Speed 4: Instant

The min/max times that speeds 1-3 cap at are to ensure that simulations
with small steps unfold slowly enough to make it worth setting a slow
speed, and, if there are too many steps, we run fast enough to complete
in a reasonable amount of time. (We don't care about "a reasonable amount
of time" at speed 0.)

Between the min and the max times, the time that a simulation will take
will increase linearly with the number of steps.

(The min running time is actually slightly more complex, as we ensure that
each speed always runs slightly faster than the one before it.)

A run of 10 steps will take

 - Speed 0: 10 seconds
 - Speed 1: 5 seconds
 - Speed 2: 2.5 seconds
 - Speed 3: 1 second
 - Speed 4: Instant

A run of 100 steps will take

 - Speed 0: 100 seconds
 - Speed 1: 33 seconds
 - Speed 2: 10 seconds
 - Speed 3: 5 seconds
 - Speed 4: Instant

A run of 1000 steps will take

 - Speed 0: 1000 seconds
 - Speed 1: 100 seconds
 - Speed 2: 50 seconds
 - Speed 3: 10 seconds
 - Speed 4: Instant

[#102294388]