Move Simulation Panel to inspector on right
This is mostly a wholesale move of the view to be opened by a new
button on the right (with a placeholder icon).

The simulation panel is entirely removed from the top toolbar with this
commit, soon a new play button will be added to the top.

The view is kept whole (thanks React for making this easy) with the
wrapper and top toggle-button removed.

The actions on the store are renamed to be descriptive
(simulationPanelExpanded rather than ExpandSimulationPanel), because
this is now just a notification (for the sliders), and actual control
of showing is now just down to the InspectorPanelView.

[#105334674]