Add new version of setting steps and duration.
(This is a large commit, as it includes the new data migrations, the
new simulation run settings, the new UI, and a bunch of updated tests.)

User now sets duration directly, and sets step units currently only as a
decoration.

Update step units to include "Step" as an option. Currently the time
unit library still supports the units conversion, but this will probably
be retired.

Add a new data migration as we now only record duration and step unit.

Progressive update on Simulation Panel view to look like Dan's mockup.
Run buttons still remain (as top panel not yet implemented).

Update a bunch of tests, and remove simulation-store-test, which only
tested some unit name functionality which we no longer use.

[#105334674]