Move diagram-only setting to simulation panel
Keeping the app-settings-store after all, as this is an app-side
setting, not a simulation-setting. Currently, this is the only
app-wide setting.

The node-value logic is slightly more complicated, as the sliders now
show is the simulation panel is open *and* we are not diagram-only,
as these are no longer functionally-equivalent.

[#105334674]