Update "old" simulation to not recursively calculate values
Previously, the "old" simulation would calculate a node's value by 
asking each inbound node for its value, which might ask their own
unbound nodes for their values, until finally reaching an "independent"
node with a defined initial value.

This commit changes the "old" simulation (run when the "new simulation"
checkbox isn't checked in the settings panel) so that we just ask each
inbound node for their previous values. This, combine with an earlier
feature to allow all nodes to set initial values, means that we can
now support closed loops.

The "old" simulation is now exactly the same as the "new" simulation
except that we don't average the node's own value when calculating
the new value.

The major change in this commit is simply removing the line

`inV = sourceNode.getCurrentValue(t)`

in simulation.coffee:42. Mixed in with that change is also a small fix
to check previousValue using `?` instead of just `||`, as we were
ignoring the previousValue when it was zero.

The rest of the commit simply tears out all the `graphIsValid`-related
code.

[#112305755]