WIP Fix undo issue
The problem is the way we are batching our title-editing undoable
actions. When we edit a title, and then drag a new node down or
create a new link, we don't correctly end the batch, so the new
node dragging (and possibly more new nodes and new links) all end
up in the same batch.

Undoing this batch not only undoes more than the user expected, but also
generally throws errors that can hang the application.

This commit adds more handlers that help us know when the user has
stopped editing the node.

(A better solution might be to use the undoable command "reducing"
solution created for CODAP for this exact same situation, but we still
need the endNodeEdit calls to properly update the UI state.)

[#113103807]