Automatically end current batch on different command
We can now optionally create a batched command that will only accept
commands with a specified name. If we then try to execute a command
with a different name, the current batch will automatically be
ended first, before executing the new command.

This allows us to be sure that a command of a different type won't
be added to a batch, even if it is executed prior to a manual
endCommandBatch call.

[#113103807]