Speed up: Create functors for default relationships
This speeds up the calculations for default relations by about 50-100x
(each step now takes 1-2% of the time to calculate).

Previously, we were using the `mathjs.eval` function, which took a
string of arbitrary text and parsed and evaluated it each time. This
was a time-consuming process, and did not seem to get JITed, so
subsequent calculations took just as long.

This became apparent when we started to want to calculate averages by
running the simulation many times. One calculations of one link took
about 0.2ms. If we had a simulation with ten links, averaged 10 times
per step, and stepped 10 times, this would be 200ms, fairly noticeable.
If we wanted more accurate averaging, or running for longer, a simulation
at fastest speed could take well over a second.

Now each default relationship returns a function based on the scalar
and the vector, and the function is plain JS and can be easily JITed.

The same 10-link simulation, averaged 10 times for 10 steps now takes
under 4ms.

[#113164487]