Run model many times per step and average
We run each model step in two stages: first we run the model several
times with the original function. This "pushes" values from parents
all the way down the descendant tree. Then we run the model several
more times, and finally assign each node the average of the values it
received in stage 2.

There are likely some optimizations that can be done here: a model
with no loops has no need of the averaging stage, and a model with
only loops has no need of the pushing stage. But for a model with
both looped and non-looped nodes, this seems necessary. In any case,
with the faster function we created in d766a00b, we can afford the
extra loops (at least while we are using the default relationships).

An improvement in the future may be to see if a node's value changes
between two steps, and, if not, no longer average it.

This commit does not yet address collectors.

Update tests. Some will not work due to collectors.

[#113164487]