Always return scaled initialValue when in SQ mode
When we are in semi-quantitative mode, we will ALWAYS return 0 and 100
for the min and max, and a value between them for the initialValue.

When we flip back and forth between SQ and Q, the proportion of the
initialValue between the min and the max should always remain the
same.

E.g.
In quantitative mode we set `min=100, max=200, initialValue="150"`
We switch to SQ mode and we see `min=0, max=100, initialValue="50"`
We move the slider to 75 and switch back into Q mode, and we see
`min=100, max=200, initialValue="175"`

The mappings are extracted into their own functions, as this will
be used be the simulation engine in the next commit.

[#113164939] [#114592289]