Scale input of node when relationship is mixed SQ and Q
If we connect a Semiquantitive node to a Quantitative node, we scale
the input before calculating the relationship as if the input were
the appropriate proportion between the output's min and max.

e.g. if we connect A at 20% to B with min 100 and max 200, it's as
if we connected a 120 input to B.

If we connect a Quantitative node to a Semiquantitative node, we scale
the input as if it were semiquantitative.

e.g. if we connect A at 120, with A's min 100 and max 200, to B, then
it's as if we connected a 20 input to B

[#114592289]