Fox inspector so it doesn't close when app settings change
Previously, when the "diagramOnly" checkbox (and the new minigraphs
checkbox) were changed, the inspector panel would close, instead
of doing the shrinking animation that existed earlier. This was
due to 883742745, and the fact that changing the app settings
would update the appView state, and thus the inspector panel
props.

This is changed to an explicit function now triggered by app-view
whenever the node selection changes, and we explicitly call a method
inside inpsector-panel-view. This seems slightly un-Reactish, but is
at least supported by React even if not ideal
(https://facebook.github.io/react/tips/expose-component-functions.html)

[#102294744] [#106499762]