Unsubscribe from all listeners on component unmount
We were seeing a number of errors in the console from components that
had unmounted still receiving `setState` calls. This was most noticeable
when the simulation panel was opened and then closed, and then the model
was run.

This was happening because any component that used one of the mixins
added a listener that would (generally) set state, and this wasn't
removed when the component was unmounted.