Small graph-view and selection cleanup
Renames obsolete term `LinkStore` with `GraphStore`, to avoid confusion.

Renames `handleClick` to `handleLinkClick` to make it clear
that this doesn't handle other clicks.

Renames GraphStores' `selectLink` to more generic `clickLink`,
as we'll be using the click for title editing soon, not just
selection.

Renames `NodeInpsection` throughout.

Switches order of events in selection-manager's `selectLink`,
so that we can go from a Node selection to a Link selection
with the inspector open and not close the inspector. (For some
reason node-to-node selection changes still close the inspector.
I haven't been able to debug that yet.)