Upgrade jsPlumb -> 2.0.2
We need to move the data parameter to the actual link target so
we can find it when we try and create a link. (Don't actually
know how it was working before).

We also remove the connector style of the end points and targets,
as this overwrites the defaults and makes a much curvier curve.
Removing them just uses the defaults.

[#115793115]