Make link-source useable while hovering over node
This is more annoying than it should be, because of the difficulty
of using a hover state while another element (the jsPlumb canvas
element) is on top.

First, the link circle is made to be always rendered, but with
opacity zero. This provides a dom element for jsPlumb to attach
the drag source to, without needing to update jsPlumb on every
hover state change.

We then add CSS to show the link circle on hover.

When we mouse over the jsPlumb target, which is directly over the
link circle, we lose our hover state. Therefore, we attach a
mouseover handler to the target, which adds a class to the circle
to stay visible.

When we start dragging away, we still want the handler to be
visible, so we attach a different class every time a new link
is created.

Finally, we want to make the circle invisible if we let go of the
link (either to attach or to abandon). This required upgrading
jsPlumb to 2.0.2, which was done in dfa1ff58491

[#115793115]