Add input field overlay to link on edit; update label
This adds a new overlay to the jsPlumb connector, a text input
field which is shown when `editingLabel` is true, and calls
`handleEditLabel` when updated (blur or return).

The `.show -> $(@).focus()` line (which is like that so that
focus is only called after the input is added to the dom)
adds an expanding animation to the input. This is unintentional
but seems fine.

[#93909054]