Remove deprecated `getDOMNode`
For the majority of times where the ref points to a build-in
DOM component, `this.refs.name` will point directly to the
DOM element, so `.getDOMNode()` is unneeded now.

For the remainder of the times, we need to use
`ReactDOM.findDOMNode(this.refs.name)` or
`ReactDOM.findDOMNode(this)`.

This will be required by 0.15, so best to do it now.

[#116530505]