Create default relationships without formula
If no formula or function is defined when a relationship is first
created (which is always the case when dragging new links), we no
longer initialize with a default formula.

Added the property `isDefined` which is not used at the moment, but
will be useful when we check if we can run the model.

Add tests.

[#115777211]