Skip links without relationship when simulating
If a node has multiple inbound links, then we calculate the new value as
if the link wasn't there at all, and calculate the average without
that connection.

If there are no valid inbound links at all, we simply treat the node
as an independent node (i.e. use its original value), but note that
we don't currently treat this as an independent node from the point
of view of sliders.

[#117059255]