Nodes now check if cycle is independent
Any node that is in a cycle that *also* has an ancestor that is
completely independent will now *not* show sliders.

Nodes in cycles will only show sliders if they have no independent
ancestors.

Note that there are still cases where we may show sliders that do not
affect the running of the model (e.g. one loop feeding into another),
but this cuts down on those cases.

Updated tests.

[#117645337]