[#121420841] Improved streaming of data between Sage and CODAP.
We found that Sage modeler was (a) issuing createCases requests to CODAP sometimes sending zero cases, and (b) updating a row in the
parent table unnecessarily. The value that was being updated was the count of the child cases, a value that could be captured with a formula. So, we change the definition of the attribute to include the formula and comment out the update altogether.

There is still a lot of unnecessary chatter between Sage and CODAP, the removal of which could improve things some more. For example, CODAP
 sends many clearUndo and clearRedo requests, that it should not. But, on a reasonable machine, CODAP can keep up with graph updates, while a
 run is executing and the slider is being manipulated.