Add basic vertical version of slider
This can be used by adding `horizontal: false` to the slider
parameters.

This is the original slider, with the black line in the
middle, turned vertical. The fancy thermometer style
comes next.

As much as possible I tried to do this using generic terms
like "length" and "thickness" instead of "width" and
"height", but of course it ended up being a bunch of
`if @props.horizontal` branches...

[#132117711]