Move link source circle above node
This required some futzing with jsPlumb to get things in the right
location. jsPlumb seems to ignore the `.elm border-top` when calculating
the source and target locations, so things end up offset. This wasn't
so bad before, but not with an even wider top border, it puts things
way off. 

Luckily switching the source anchor to "Bottom" moves the box to the
right location, but that's a bit of a coincidence, and if we move things
around further this offset may not always work. We may need to dive into
jsPlumb to get it to calculate the element's location correctly.