Fix initial render bug
In certain loading situations, the graph view is mounted after the
graph store data has been loaded.

(View mounting is asynchronous, and we should assume it could happen at
any time.)

When the graph view is already mounted, it has a listener to the
GraphStore which updates the view's state when the graph store receives
data. This works fine.

However, if the view was not mounted, it will not receive any events,
and will need to rely on `getInitialState` to provide its state. This
in theory is fine, but there was a bug in the mixin's `getInitialState`.

Since `getInitialState` is only actually relevant in this async
code-path, this wasn't noticed.

Note that we should not really be using mixing at all with React 0.14,
but they currently seem to work as expected.
https://medium.com/@dan_abramov/mixins-are-dead-long-live-higher-order-components-94a0d2f9e750#.ey4zm3ngk

[#133127059]