Replace jQuery slider with pure React
The API for this slider remains the same, except for `horizontal` being
replaced by `orientation`, which allows for an easier look-up function.

Although we currently only have the one use of the slider (the nodes),
the various options still work, such as horizontal sliders and discrete
step sizes. See /test.html for the demos.

[#133192365]