Get correct step units when we have collectors
This changes `simulation-store` to pass back the user-selected units
whenever there is a collector in he graph, and not just while running
a period-simulation, as before.

The only effect of this should be that, when we click the Graph button
on a node, we will correctly send CODAP the correct units for the
time axis of the graph.

[#133192489]