Run synchronously on runSimulation action
Previously, if we called `runSimulation` many times in succession (e.g.
when moving the slider quickly), we would batch up all the calls and
then run them all at once.

This would lead to running the simulation multiple times for the same
value, because when we run the simulation we only look at the current
value (the value isn't cached when we call the action).

This change makes the `runSimulation` action run synchronously.

This seems to have a positive effect on the feeling of "lag" when
running the simulation.