Compare links to decide whether to redraw jsPlumb
This introduces a graph "description" which contains easily-comparable
descriptions of (eventually) different aspects of the graph, to allow
us to make better decisions about whether to run updates.

In this case, `description.links` is a string description of all the
links' terminal locations and relationship formulas (for stroke styles).

When the graph's state updates, we can compare the previous description
with the current one to decide wether to redraw the jsPlumb diagram.

This speeds up the running of simulations considerably, because we were
previously spending lots of cycles redrawing jsPlumb every time we
updated the nodes by running a step.

Add test for description

[#134467689]