Explicitly refer to data context when creating cases [#137216155]
At init time we call check for the CODAP dataContext, "Sage Modeler", but
when creating cases and managing cases we were not. CODAP was getting
confused about the default data context for Sage, and so, was applying
the addition to the wrong one. Modified all resource definitions in
codap-connect.coffee to remove the ambiguity.