[#141349005] Fix bug: Graph button does not open a graph in CODAP
There were three places _createMissingDataAttributes was called in CodapConnect.
Two did not provide a callback. Logic internal to that method was assuming a callback.
It did not appear that a callback was required in the two places it wasn't provided,
so I added a test for null callback to the method.