Fixed clicking outside of link label editor losing change [#141320637]
Previously the code was using the currently selected link when the blur handler on the link input was invoked due to a click outside of the link input.  However, the click would cause the currently selected link to be set to null before the blur handler was called causing the link not to be updated.

The code now adds the link to the closure that sets up the editing overlay so that it is available when the blur event is invoked.