Make deleting a node more intuitive when variable name editing is the focus. [#136380649]
With this change the code now tracks if a node is added during the session (as opposed to loaded from a save) and then deletes the node if during the initial edit the first key pressed is either backspace or delete.

This also fixes the global key handler to also listen for delete (46) along with the existing backspace (8) listener.