Nodes must have unique names [#134798657]
Adds a check when either a file is loaded or when a title is edited to require both a unique node name and the non-default node name ("Untitled"). If the name is not unique a number is appended to it starting at either 2 if the existing node name does not end with an integer or the ending integer + 1.

For visual feedback during title editing the input turns red when the name is not unique.