Fix undo/redo buttons to work in standalone mode [#133193161]
With this change Sage sends undoButtonPress/redoButtonPress to CODAP when in standalone mode instead of undoAction/redoAction.  This allows CODAP to control the undo/redo stack while still allowing the undo/redo buttons to be usable in Sage.

NOTE: there seems to be an issue with CODAP not sending undoableActionPeformed messages down to Sage when things like attributes are added to datatables.  This can cause the Sage undo/redo stack to get out of order.  I'm pushing this commit so we have a deployed branch to test CODAP against.