Changed default recording unit to be days [#135136847]
The unit is switched to days (or the saved or last selected collector unit) internally now when the graph transitions from 0 to 1 collector nodes.  It switches back to steps when the graph transitions from 1 to 0 collector nodes.  Otherwise any collector change keeps keeps the units as they are (since a user may then change the units with the simulation panel dropdown)

This also fixes a bug where the time unit name setting in the simulation store was not being set when data was imported, only the time unit itself was being set.  This caused the dropdown in the simulation panel to show an translation error.