Fixed undo/redo buttons so they work in standalone mode [#133193161]
This change disables sending any events to CODAP about undo/redo when we are in standalone mode and instead relies on the Sage undo/redo manager to handle the requsests.

Before this change when standalone mode was detected the Sage undo/redo actions would be skipped and instead CODAP would be informed of the need for undo/redo. However this effectivly disabled the undo/redo buttons showing in Sage.

A new graph-store variable that tracks if we are running in CODAP (set via codap-connect) is now used to flag when undo/redo events should be sent to CODAP.

NOTE: there needs to be work done in CODAP in the DI API so that CODAP can notify the DI the state of the undo/redo stack.