Migration: "diagramOnly" -> "complexity"
This removes the `diagramOnly` property from the AppSettingsStore, and
replaces it with the property `complexity`, which can be one of several
string values.

This adds a migration for data, which will default to adding the prop
`complexity: "basic"` to the user state, unless the user already had
`diagramOnly: true`, in which case we'll set complexity to `diagramOnly`

Update tests.

Note: this is a BREAKING CHANGE until the next commit, which replaces
all references to `AppSettingsStore.diagramOnly` with `complexity`
(commits separated for clarity).