Replace all references to diagramOnly
We now check if 

`complexity is AppSettingsStore.store.Complexity.diagramOnly`

which is rather long-winded, but seems necessary as we're going to have
a number of fine-grained references to `complexity` now.

Call `setComplexity` instead of `diagramOnly`.

App's behavior is now identical to previous behavior, except we now have
radio buttons instead of the checkbox.