Simplified link pulldowns for basic relationships
If we have a "basic" complexity, we

1. Only show "increase" and "decrease" options in the vector pulldown
2. Auto-set scalar to "about the same"
3. Hide the scalar pulldown

This results in the relationship being set instantly when "increase" or
"decrease" are selected, and the graph being drawn.

If the user then selects the "expanded" or "collectors" complexity, the
pulldowns will reflect the state as if the user had selected the "about
the same" scalar pulldown, and the the rest of the scalar options, and
the "varies" option, will appear.