Prevent selection of lower complexity; +tests
Once a graph has certain features, the user cannot lower the complexity
of the diagram to a setting that forbids those features.

We analyze the current graph and detect the minimum complexity present
in the graph. We then disable the radio button choices for those lower
complexities.

If the user removes those features from the graph, the choices are
re-enabled.

Add tests for the minimum-complexity calculation.