New set-initial-value relation type for collectors
This adds a new relation type, `initial-value`. This can only be used
from a non-accumulator to an accumulator. This will simply set the
initial value of a collector node before the simulation run starts.

Style the link like a custom data link (thick gray).

For the pull-down menu, we need to change the wording slightly.
We currently have hard-coded "a IS [added to] b EACH STEP". The best
wording for the new pull-down is "a [sets the initial value of] b".
As a slight hack, this relation simply has a setting that tells us to
hide the "IS ... EACH STEP" text. Eventually, relations may need
finer-grained control over the surrounding text.